Refactor primitive pointer tests (#5425)
